requirements.txt
streamlit==1.32.0
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
python-dotenv==1.0.1
google-ads==22.1.0 